fix(frontend): move column types modal inside edit dropdown
